# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DE — Desert Emerald API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SearchInput {
  /** Search query in Arabic or English */
  query: string;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  region?: string | null;
  /**
     * Filter by category e.g. Medical, صيدلية, General_Business
     * @nullable
     */
  category?: string | null;
  /** @nullable */
  limit?: number | null;
}

export interface Entity {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  governorate?: string | null;
  /** @nullable */
  doctor_name?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  source?: string | null;
}

export interface SearchResponse {
  results: Entity[];
  total: number;
  query: string;
  /**
     * AI-style summary answer
     * @nullable
     */
  answer?: string | null;
}

export interface StatItem {
  name: string;
  count: number;
}

export interface StatsResponse {
  total_entities: number;
  total_with_phone: number;
  total_verified: number;
  top_specialties: StatItem[];
  top_regions: StatItem[];
  category_breakdown: StatItem[];
}

export interface Suggestion {
  text: string;
  type: string;
}

export type DeSuggestParams = {
q: string;
};