# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DE — Desert Emerald API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /de/search:
    post:
      operationId: deSearch
      tags: [de]
      summary: Search DE database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchInput"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
...
components:
  schemas:
...
    SearchInput:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Search query in Arabic or English
        specialty:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
          description: Filter by category e.g. Medical, صيدلية, General_Business
        limit:
          type: ["integer", "null"]
    Entity:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        governorate:
          type: ["string", "null"]
        doctor_name:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        facebook:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
    SearchResponse:
      type: object
      required:
        - results
        - total
        - query
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        total:
          type: integer
        query:
          type: string
        answer:
          type: ["string", "null"]
          description: AI-style summary answer
...