# lib/api-zod/src/generated/types/entity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DE — Desert Emerald API
 * OpenAPI spec version: 0.1.0
 */

export interface Entity {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  governorate?: string | null;
  /** @nullable */
  doctor_name?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  source?: string | null;
}