# lib/api-zod/src/generated/types/searchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DE — Desert Emerald API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchInput {
  /** Search query in Arabic or English */
  query: string;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  region?: string | null;
  /**
     * Filter by category e.g. Medical, صيدلية, General_Business
     * @nullable
     */
  category?: string | null;
  /** @nullable */
  limit?: number | null;
}