# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DE — Desert Emerald API
 * OpenAPI spec version: 0.1.0
 */
import type { Entity } from './entity';

export interface SearchResponse {
  results: Entity[];
  total: number;
  query: string;
  /**
     * AI-style summary answer
     * @nullable
     */
  answer?: string | null;
}